+++ src/repositories/user/user.repository.spec.ts
@@ -1,6 +1,6 @@
 import { Test } from '@nestjs/testing';
 import { User, UserCredentials, UserProps } from 'src/entities/user.entity';
-import { UserDataSource } from './user.datasource';
+import { UserRepository } from './user.repository';
 import { MongooseModule } from '@nestjs/mongoose';
 import { UserModel, UserSchema } from './user.schema';
 import { ConfigModule } from '@nestjs/config';
@@ -9,8 +9,8 @@ import { Model } from 'mongoose';
 import { HttpException, HttpStatus } from '@nestjs/common';
 import * as bcrypt from 'bcrypt';
 
-describe('UserDataSource', () => {
-  let userDataSource: UserDataSource;
+describe('UserRepository', () => {
+  let userRepository: UserRepository;
   let userModel: Model<UserModel>;
 
   const password = 'Strong123!';
@@ -35,11 +35,11 @@ describe('UserDataSource', () => {
           },
         ]),
       ],
-      providers: [UserDataSource],
+      providers: [UserRepository],
     }).compile();
 
-    userDataSource = testModule.get<UserDataSource>(UserDataSource);
-    userModel = (userDataSource as any).userModel;
+    userRepository = testModule.get<UserRepository>(UserRepository);
+    userModel = (userRepository as any).userModel;
   });
 
   afterEach(async () => {
@@ -48,7 +48,7 @@ describe('UserDataSource', () => {
 
   describe('create', () => {
     it('should create user in db', async () => {
-      const user = await userDataSource.create(userProps);
+      const user = await userRepository.create(userProps);
 
       expect(user.constructor.name).toBe(User.name);
       expect(isUUID(user.id)).toBeTruthy;
@@ -63,7 +63,7 @@ describe('UserDataSource', () => {
     it('should throw an error if user email is already taken', async () => {
       await userModel.create(userProps);
 
-      await expect(userDataSource.create(userProps)).rejects.toEqual(
+      await expect(userRepository.create(userProps)).rejects.toEqual(
         new HttpException('User already exists', HttpStatus.CONFLICT),
       );
     });
@@ -78,7 +78,7 @@ describe('UserDataSource', () => {
     it('should find user with valid credentials', async () => {
       await userModel.create(userProps);
 
-      const user = await userDataSource.findOneByCredentials(credentials);
+      const user = await userRepository.findOneByCredentials(credentials);
 
       expect(user.constructor.name).toBe(User.name);
       expect(isUUID(user.id)).toBeTruthy;
@@ -94,7 +94,7 @@ describe('UserDataSource', () => {
       await userModel.create(userProps);
 
       await expect(
-        userDataSource.findOneByCredentials({
+        userRepository.findOneByCredentials({
           ...credentials,
           email: 'wrong@email.com',
         }),
@@ -107,7 +107,7 @@ describe('UserDataSource', () => {
       await userModel.create(userProps);
 
       await expect(
-        userDataSource.findOneByCredentials({
+        userRepository.findOneByCredentials({
           ...credentials,
           password: 'wrongPassword',
         }),
